# backend/openapi.yaml
openapi: 3.1.0
info:
  title: RMG ERP System
  version: 1.0.0
  description: |
    Southern Apparels ERP System API
    Multi-Database Architecture with 5 separate PostgreSQL databases
    - Clients DB: Buyers, Suppliers, Contacts, Shipping, Banking
    - Samples DB: Styles, Variants, Samples, TNA, Plans, Operations, Materials, Colors, Sizes
    - Users DB: User management & Authentication
    - Orders DB: Order management
    - Merchandiser DB: Future merchandiser features
servers:
  - url: /
    description: Current server (auto-detected)
...
paths:
...
  /api/v1/master/sizes:
    get:
      tags: [master-data]
      summary: List all sizes
      responses:
        '200':
          description: List of sizes
          content:
            application/json:
              example:
                - id: 1
                  size_value: "S"
                  size_label: "Small"
                  size_category: "Adult"
                  sort_order: 1
                  is_active: true
                  created_at: "2024-01-01T00:00:00Z"
                - id: 2
                  size_value: "M"
                  size_label: "Medium"
                  size_category: "Adult"
                  sort_order: 2
                  is_active: true
                - id: 3
                  size_value: "L"
                  size_label: "Large"
                  size_category: "Adult"
                  sort_order: 3
                  is_active: true
    post:
      tags: [master-data]
      summary: Create size
      requestBody:
        required: true
        content:
          application/json:
            example:
              size_value: "XXL"
              size_label: "Extra Extra Large"
              size_category: "Adult"
              sort_order: 5
              is_active: true
      responses:
        '201':
          description: Size created
  /api/v1/master/sizes/{size_id}:
    get:
      tags: [master-data]
      summary: Get size by ID
      parameters:
        - name: size_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Size details
        '404': { description: Not found }
    put:
      tags: [master-data]
      summary: Update size
      parameters:
        - name: size_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            example:
              size_label: "Small (US)"
              is_active: true
      responses:
        '200':
          description: Size updated
    delete:
      tags: [master-data]
      summary: Delete size
      parameters:
        - name: size_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204': { description: Deleted }

  # ============ MERCHANDISER ============
...